+++ src/main/java/com/example/aboutme/Login/controller/GoogleController.java
@@ -2,8 +2,7 @@ package com.example.aboutme.Login.controller;
 
 import com.example.aboutme.Login.common.MsgEntity;
 import com.example.aboutme.Login.dto.GoogleDTO;
-import com.example.aboutme.Login.dto.KakaoDTO;
-import com.example.aboutme.Login.service.GoogleService;
+import com.example.aboutme.Login.service.GoogleServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -16,7 +15,7 @@ import javax.servlet.http.HttpServletRequest;
 @RequiredArgsConstructor
 @RequestMapping("google")
 public class GoogleController {
-    private final GoogleService googleService;
+    private final GoogleServiceImpl googleService;
 
     @GetMapping("/callback")
     public ResponseEntity<MsgEntity> callback(HttpServletRequest request) throws Exception {
