+++ src/main/java/com/example/aboutme/Login/service/KakaoServiceImpl.java
@@ -1,0 +1,121 @@
+package com.example.aboutme.Login.service;
+
+import com.example.aboutme.Login.MemberConverter;
+import com.example.aboutme.Login.dto.KakaoDTO;
+import com.example.aboutme.domain.Member;
+import com.example.aboutme.domain.constant.Social;
+import com.example.aboutme.repository.MemberRepository;
+import lombok.RequiredArgsConstructor;
+import org.json.simple.JSONObject;
+import org.json.simple.parser.JSONParser;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Service;
+import org.springframework.util.LinkedMultiValueMap;
+import org.springframework.util.MultiValueMap;
+import org.springframework.web.client.RestTemplate;
+
+@Service
+@RequiredArgsConstructor
+public class KakaoServiceImpl implements KakaoService{
+    private final MemberRepository memberRepository;
+
+    @Value("${spring.security.oauth2.client.registration.kakao.client-id}")
+    private String KAKAO_CLIENT_ID;
+
+    @Value("${spring.security.oauth2.client.registration.kakao.client-secret}")
+    private String KAKAO_CLIENT_SECRET;
+
+    @Value("${spring.security.oauth2.client.registration.kakao.redirect-uri}")
+    private String KAKAO_REDIRECT_URL;
+
+    private final static String KAKAO_AUTH_URI = "https://kauth.kakao.com";
+    private final static String KAKAO_API_URI = "https://kapi.kakao.com";
+
+    public String getKakaoLogin() {
+        return KAKAO_AUTH_URI + "/oauth/authorize"
+                + "?client_id=" + KAKAO_CLIENT_ID
+                + "&redirect_uri=" + KAKAO_REDIRECT_URL
+                + "&response_type=code";
+    }
+
+    public KakaoDTO getKakaoInfo(String code) throws Exception {
+        if (code == null) throw new Exception("Failed get authorization code");
+
+        String accessToken = "";
+        String refreshToken = "";
+
+        try {
+            HttpHeaders headers = new HttpHeaders();
+            headers.add("Content-type", "application/x-www-form-urlencoded");
+
+            MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
+            params.add("grant_type"   , "authorization_code");
+            params.add("client_id"    , KAKAO_CLIENT_ID);
+            params.add("client_secret", KAKAO_CLIENT_SECRET);
+            params.add("code"         , code);
+            params.add("redirect_uri" , KAKAO_REDIRECT_URL);
+
+            RestTemplate restTemplate = new RestTemplate();
+            HttpEntity<MultiValueMap<String, String>> httpEntity = new HttpEntity<>(params, headers);
+
+            ResponseEntity<String> response = restTemplate.exchange(
+                    KAKAO_AUTH_URI + "/oauth/token",
+                    HttpMethod.POST,
+                    httpEntity,
+                    String.class
+            );
+
+            JSONParser jsonParser = new JSONParser();
+            JSONObject jsonObj = (JSONObject) jsonParser.parse(response.getBody());
+
+            accessToken  = (String) jsonObj.get("access_token");
+            refreshToken = (String) jsonObj.get("refresh_token");
+        } catch (Exception e) {
+            throw new Exception("API call failed");
+        }
+
+        return getUserInfoWithToken(accessToken);
+    }
+
+    public KakaoDTO getUserInfoWithToken(String accessToken) throws Exception {
+        //HttpHeader 생성
+        HttpHeaders headers = new HttpHeaders();
+        headers.add("Authorization", "Bearer " + accessToken);
+        headers.add("Content-type", "application/x-www-form-urlencoded;charset=utf-8");
+
+        //HttpHeader 담기
+        RestTemplate rt = new RestTemplate();
+        HttpEntity<MultiValueMap<String, String>> httpEntity = new HttpEntity<>(headers);
+        ResponseEntity<String> response = rt.exchange(
+                KAKAO_API_URI + "/v2/user/me",
+                HttpMethod.GET,
+                httpEntity,
+                String.class
+        );
+
+        //Response 데이터 파싱
+        JSONParser jsonParser = new JSONParser();
+        JSONObject jsonObj    = (JSONObject) jsonParser.parse(response.getBody());
+        JSONObject account = (JSONObject) jsonObj.get("kakao_account");
+        JSONObject profile = (JSONObject) account.get("profile");
+
+        long id = (long) jsonObj.get("id");
+        System.out.println(id);
+        String email = String.valueOf(account.get("email"));
+        String nickname = String.valueOf(profile.get("nickname"));
+
+        return KakaoDTO.builder()
+                .id(id)
+                .email(email)
+                .nickname(nickname).build();
+    }
+
+    public void saveKakaoMember(KakaoDTO kakaoDTO){
+        Member newMember = MemberConverter.toMember(kakaoDTO, Social.KAKAO);
+        memberRepository.save(newMember);
+    }
+}
